fix: silence certian logs
