+++ packages/backend/src/app/app-event-routing/app-event-routing.module.ts
@@ -35,7 +35,7 @@ export const appEventRoutingController = async (fastify: FastifyInstance) => {
             config: {
                 rawBody: true,
             },
-            logLevel: 'debug',
+            logLevel: 'silent',
         },
         async (
             request: FastifyRequest<{
